# Switch allocator to mimalloc for the DB CLI