--- a/Cargo.toml
+++ b/Cargo.toml
@@ -115,6 +115,7 @@ itertools = "0.10.5"
 jsonwebtoken = { version = "8.1.0" }
 lazy_static = "1.4.0"
 log = "0.4.17"
+mimalloc = "0.1.39"
 nonempty = "0.8.1"
 once_cell = "1.16"
 parking_lot = { version = "0.12.1", features = ["send_guard", "arc_lock"] }