# Description of Changes

We are doing a lot of allocations / clones, and I figured it might be worth looking into high-performance allocators as an easy win among other things.

mimalloc generally beats other allocators and would make the biggest difference in multi-threaded environment (such as an actual cloud) as it has separate per-thread pools.

However, it makes a noticeable difference even in our single-threaded benchmarks:

<details><summary>Benchmark results</summary>

# Benchmark Report

Legend:

- `load`: number of rows pre-loaded into the database
- `count`: number of rows touched by the transaction
- index types:
    - `unique`: a single index on the `id` column
    - `non_unique`: no indexes
    - `multi_index`: non-unique index on every column
- schemas:
    - `person(id: u32, name: String, age: u64)`
    - `location(id: u32, x: u64, y: u64)`

All throughputs are single-threaded.

    
## Empty transaction

| db          | on disk | new latency   | old latency   | new throughput | old throughput | 
|-------------|---------|---------------|---------------|----------------|----------------|
| sqlite      | 💿       | 437.7±1.19ns  | 458.1±8.77ns  | -              | -              | 
| sqlite      | 🧠       | 429.9±1.77ns  | 457.5±6.57ns  | -              | -              | 
| stdb_module | 💿       | 15.9±0.36µs   | 15.6±0.36µs   | -              | -              | 
| stdb_module | 🧠       | 16.3±0.48µs   | 15.9±0.44µs   | -              | -              | 
| stdb_raw    | 💿       | 104.4±0.21ns  | 103.7±0.12ns  | -              | -              | 
| stdb_raw    | 🧠       | 104.2±0.29ns  | 103.4±0.13ns  | -              | -              | 

## Single-row insertions

| db          | on disk | schema   | index type  | load | new latency     | old latency      | new throughput | old throughput | 
|-------------|---------|----------|-------------|------|-----------------|------------------|----------------|----------------|
| sqlite      | 💿       | location | multi_index | 0    | 14.7±1.15µs     | 14.6±0.07µs      | 66.5 Ktx/sec   | 66.7 Ktx/sec   | 
| sqlite      | 💿       | location | multi_index | 1000 | 15.8±0.11µs     | 16.0±0.13µs      | 62.0 Ktx/sec   | 60.9 Ktx/sec   | 
| sqlite      | 💿       | location | non_unique  | 0    | 7.2±1.26µs      | 7.4±0.04µs       | 135.8 Ktx/sec  | 132.8 Ktx/sec  | 
| sqlite      | 💿       | location | non_unique  | 1000 | 6.9±0.04µs      | 7.2±0.04µs       | 140.9 Ktx/sec  | 135.9 Ktx/sec  | 
| sqlite      | 💿       | location | unique      | 0    | 7.0±0.04µs      | 7.4±0.33µs       | 139.4 Ktx/sec  | 131.5 Ktx/sec  | 
| sqlite      | 💿       | location | unique      | 1000 | 7.0±0.02µs      | 7.2±0.04µs       | 140.0 Ktx/sec  | 135.8 Ktx/sec  | 
| sqlite      | 💿       | person   | multi_index | 0    | 14.2±0.03µs     | 14.7±0.71µs      | 69.0 Ktx/sec   | 66.5 Ktx/sec   | 
| sqlite      | 💿       | person   | multi_index | 1000 | 16.1±0.18µs     | 16.2±0.13µs      | 60.8 Ktx/sec   | 60.3 Ktx/sec   | 
| sqlite      | 💿       | person   | non_unique  | 0    | 7.2±0.02µs      | 7.5±0.35µs       | 135.9 Ktx/sec  | 130.5 Ktx/sec  | 
| sqlite      | 💿       | person   | non_unique  | 1000 | 7.2±0.05µs      | 7.4±0.04µs       | 135.3 Ktx/sec  | 132.2 Ktx/sec  | 
| sqlite      | 💿       | person   | unique      | 0    | 7.2±0.41µs      | 7.4±0.33µs       | 135.2 Ktx/sec  | 131.6 Ktx/sec  | 
| sqlite      | 💿       | person   | unique      | 1000 | 7.2±0.05µs      | 7.4±0.05µs       | 135.2 Ktx/sec  | 132.1 Ktx/sec  | 
| sqlite      | 🧠       | location | multi_index | 0    | 4.0±0.02µs      | 4.0±0.01µs       | 242.7 Ktx/sec  | 241.3 Ktx/sec  | 
| sqlite      | 🧠       | location | multi_index | 1000 | 5.1±0.04µs      | 5.2±0.03µs       | 191.5 Ktx/sec  | 186.5 Ktx/sec  | 
| sqlite      | 🧠       | location | non_unique  | 0    | 1853.8±3.31ns   | 1888.8±7.69ns    | 526.8 Ktx/sec  | 517.0 Ktx/sec  | 
| sqlite      | 🧠       | location | non_unique  | 1000 | 1907.5±11.85ns  | 1910.4±9.81ns    | 512.0 Ktx/sec  | 511.2 Ktx/sec  | 
| sqlite      | 🧠       | location | unique      | 0    | 1840.3±4.62ns   | 1880.8±5.27ns    | 530.7 Ktx/sec  | 519.2 Ktx/sec  | 
| sqlite      | 🧠       | location | unique      | 1000 | 1960.9±10.60ns  | 1951.8±7.85ns    | 498.0 Ktx/sec  | 500.3 Ktx/sec  | 
| sqlite      | 🧠       | person   | multi_index | 0    | 3.6±0.01µs      | 4.0±0.01µs       | 268.2 Ktx/sec  | 243.3 Ktx/sec  | 
| sqlite      | 🧠       | person   | multi_index | 1000 | 5.4±0.02µs      | 5.7±0.03µs       | 179.6 Ktx/sec  | 171.4 Ktx/sec  | 
| sqlite      | 🧠       | person   | non_unique  | 0    | 1947.7±7.26ns   | 1970.5±6.36ns    | 501.4 Ktx/sec  | 495.6 Ktx/sec  | 
| sqlite      | 🧠       | person   | non_unique  | 1000 | 2.0±0.01µs      | 2.1±0.01µs       | 485.0 Ktx/sec  | 465.7 Ktx/sec  | 
| sqlite      | 🧠       | person   | unique      | 0    | 1932.6±5.80ns   | 1951.3±5.15ns    | 505.3 Ktx/sec  | 500.5 Ktx/sec  | 
| sqlite      | 🧠       | person   | unique      | 1000 | 2.1±0.02µs      | 2.1±0.01µs       | 471.3 Ktx/sec  | 456.1 Ktx/sec  | 
| stdb_module | 💿       | location | multi_index | 0    | 40.5±4.36µs     | 35.1±4.62µs      | 24.1 Ktx/sec   | 27.9 Ktx/sec   | 
| stdb_module | 💿       | location | multi_index | 1000 | 124.2±8.34µs    | 136.0±55.50µs    | 7.9 Ktx/sec    | 7.2 Ktx/sec    | 
| stdb_module | 💿       | location | non_unique  | 0    | 29.6±2.73µs     | 35.0±3.03µs      | 33.0 Ktx/sec   | 27.9 Ktx/sec   | 
| stdb_module | 💿       | location | non_unique  | 1000 | 125.8±26.80µs   | 230.7±3.00µs     | 7.8 Ktx/sec    | 4.2 Ktx/sec    | 
| stdb_module | 💿       | location | unique      | 0    | 34.8±3.30µs     | 40.4±4.30µs      | 28.0 Ktx/sec   | 24.2 Ktx/sec   | 
| stdb_module | 💿       | location | unique      | 1000 | 266.2±8.40µs    | 383.8±8.09µs     | 3.7 Ktx/sec    | 2.5 Ktx/sec    | 
| stdb_module | 💿       | person   | multi_index | 0    | 48.6±5.69µs     | 52.7±6.65µs      | 20.1 Ktx/sec   | 18.5 Ktx/sec   | 
| stdb_module | 💿       | person   | multi_index | 1000 | 307.6±11.96µs   | 471.2±109.22µs   | 3.2 Ktx/sec    | 2.1 Ktx/sec    | 
| stdb_module | 💿       | person   | non_unique  | 0    | 35.2±3.55µs     | 35.8±3.54µs      | 27.7 Ktx/sec   | 27.3 Ktx/sec   | 
| stdb_module | 💿       | person   | non_unique  | 1000 | 265.7±60.75µs   | 259.0±113.00µs   | 3.7 Ktx/sec    | 3.8 Ktx/sec    | 
| stdb_module | 💿       | person   | unique      | 0    | 38.7±3.94µs     | 44.0±4.17µs      | 25.2 Ktx/sec   | 22.2 Ktx/sec   | 
| stdb_module | 💿       | person   | unique      | 1000 | 352.7±12.98µs   | 435.2±8.06µs     | 2.8 Ktx/sec    | 2.2 Ktx/sec    | 
| stdb_module | 🧠       | location | multi_index | 0    | 32.7±2.86µs     | 30.1±2.64µs      | 29.8 Ktx/sec   | 32.4 Ktx/sec   | 
| stdb_module | 🧠       | location | multi_index | 1000 | 208.7±82.66µs   | 217.6±15.56µs    | 4.7 Ktx/sec    | 4.5 Ktx/sec    | 
| stdb_module | 🧠       | location | non_unique  | 0    | 27.7±2.06µs     | 26.0±1.45µs      | 35.3 Ktx/sec   | 37.6 Ktx/sec   | 
| stdb_module | 🧠       | location | non_unique  | 1000 | 112.1±3.87µs    | 230.1±18.37µs    | 8.7 Ktx/sec    | 4.2 Ktx/sec    | 
| stdb_module | 🧠       | location | unique      | 0    | 31.3±2.35µs     | 28.9±1.86µs      | 31.2 Ktx/sec   | 33.8 Ktx/sec   | 
| stdb_module | 🧠       | location | unique      | 1000 | 109.9±41.41µs   | 304.6±20.93µs    | 8.9 Ktx/sec    | 3.2 Ktx/sec    | 
| stdb_module | 🧠       | person   | multi_index | 0    | 40.3±3.13µs     | 38.1±3.17µs      | 24.2 Ktx/sec   | 25.6 Ktx/sec   | 
| stdb_module | 🧠       | person   | multi_index | 1000 | 151.5±18.64µs   | 210.5±49.11µs    | 6.4 Ktx/sec    | 4.6 Ktx/sec    | 
| stdb_module | 🧠       | person   | non_unique  | 0    | 30.4±2.98µs     | 28.4±2.90µs      | 32.2 Ktx/sec   | 34.4 Ktx/sec   | 
| stdb_module | 🧠       | person   | non_unique  | 1000 | 152.9±37.22µs   | 247.1±21.15µs    | 6.4 Ktx/sec    | 4.0 Ktx/sec    | 
| stdb_module | 🧠       | person   | unique      | 0    | 31.2±2.42µs     | 34.7±3.07µs      | 31.3 Ktx/sec   | 28.2 Ktx/sec   | 
| stdb_module | 🧠       | person   | unique      | 1000 | 229.6±33.70µs   | 385.1±14.01µs    | 4.3 Ktx/sec    | 2.5 Ktx/sec    | 
| stdb_raw    | 💿       | location | multi_index | 0    | 5.0±0.01µs      | 5.3±0.01µs       | 196.6 Ktx/sec  | 184.3 Ktx/sec  | 
| stdb_raw    | 💿       | location | multi_index | 1000 | 27.9±206.60µs   | 33.8±0.51µs      | 34.9 Ktx/sec   | 28.9 Ktx/sec   | 
| stdb_raw    | 💿       | location | non_unique  | 0    | 3.4±0.01µs      | 3.5±0.01µs       | 290.9 Ktx/sec  | 277.3 Ktx/sec  | 
| stdb_raw    | 💿       | location | non_unique  | 1000 | 4.7±0.29µs      | 29.7±79.92µs     | 207.4 Ktx/sec  | 32.8 Ktx/sec   | 
| stdb_raw    | 💿       | location | unique      | 0    | 4.1±0.01µs      | 4.3±0.01µs       | 239.2 Ktx/sec  | 224.8 Ktx/sec  | 
| stdb_raw    | 💿       | location | unique      | 1000 | 6.2±0.16µs      | 39.9±127.80µs    | 157.3 Ktx/sec  | 24.5 Ktx/sec   | 
| stdb_raw    | 💿       | person   | multi_index | 0    | 8.3±0.01µs      | 9.0±0.01µs       | 117.4 Ktx/sec  | 108.7 Ktx/sec  | 
| stdb_raw    | 💿       | person   | multi_index | 1000 | 53.7±422.91µs   | 59.6±350.92µs    | 18.2 Ktx/sec   | 16.4 Ktx/sec   | 
| stdb_raw    | 💿       | person   | non_unique  | 0    | 3.9±0.01µs      | 4.1±0.04µs       | 248.7 Ktx/sec  | 236.1 Ktx/sec  | 
| stdb_raw    | 💿       | person   | non_unique  | 1000 | 5.5±0.08µs      | 14.2±0.13µs      | 176.9 Ktx/sec  | 68.8 Ktx/sec   | 
| stdb_raw    | 💿       | person   | unique      | 0    | 5.6±0.03µs      | 5.9±0.01µs       | 174.8 Ktx/sec  | 165.5 Ktx/sec  | 
| stdb_raw    | 💿       | person   | unique      | 1000 | 8.2±0.14µs      | 21.7±0.35µs      | 119.6 Ktx/sec  | 45.1 Ktx/sec   | 
| stdb_raw    | 🧠       | location | multi_index | 0    | 3.6±0.01µs      | 4.1±0.01µs       | 273.8 Ktx/sec  | 237.6 Ktx/sec  | 
| stdb_raw    | 🧠       | location | multi_index | 1000 | 5.2±0.06µs      | 29.4±0.79µs      | 188.1 Ktx/sec  | 33.2 Ktx/sec   | 
| stdb_raw    | 🧠       | location | non_unique  | 0    | 2.1±0.01µs      | 2.2±0.01µs       | 475.0 Ktx/sec  | 441.5 Ktx/sec  | 
| stdb_raw    | 🧠       | location | non_unique  | 1000 | 2.7±0.01µs      | 19.4±0.08µs      | 364.4 Ktx/sec  | 50.4 Ktx/sec   | 
| stdb_raw    | 🧠       | location | unique      | 0    | 2.7±0.01µs      | 3.0±0.01µs       | 356.5 Ktx/sec  | 322.7 Ktx/sec  | 
| stdb_raw    | 🧠       | location | unique      | 1000 | 4.0±0.04µs      | 24.5±0.12µs      | 242.8 Ktx/sec  | 39.9 Ktx/sec   | 
| stdb_raw    | 🧠       | person   | multi_index | 0    | 7.0±0.02µs      | 7.6±0.01µs       | 140.0 Ktx/sec  | 128.6 Ktx/sec  | 
| stdb_raw    | 🧠       | person   | multi_index | 1000 | 9.1±0.06µs      | 21.5±1.16µs      | 107.7 Ktx/sec  | 45.4 Ktx/sec   | 
| stdb_raw    | 🧠       | person   | non_unique  | 0    | 2.6±0.00µs      | 2.8±0.01µs       | 374.7 Ktx/sec  | 345.8 Ktx/sec  | 
| stdb_raw    | 🧠       | person   | non_unique  | 1000 | 3.4±0.02µs      | 12.8±0.17µs      | 286.2 Ktx/sec  | 76.2 Ktx/sec   | 
| stdb_raw    | 🧠       | person   | unique      | 0    | 4.3±0.00µs      | 4.7±0.01µs       | 227.7 Ktx/sec  | 209.3 Ktx/sec  | 
| stdb_raw    | 🧠       | person   | unique      | 1000 | 5.7±0.05µs      | 17.0±0.57µs      | 170.3 Ktx/sec  | 57.5 Ktx/sec   | 

## Multi-row insertions

| db          | on disk | schema   | index type  | load | count | new latency       | old latency       | new throughput | old throughput | 
|-------------|---------|----------|-------------|------|-------|-------------------|-------------------|----------------|----------------|
| sqlite      | 💿       | location | multi_index | 0    | 100   | 128.6±3.30µs      | 132.9±2.68µs      | 7.6 Ktx/sec    | 7.3 Ktx/sec    | 
| sqlite      | 💿       | location | multi_index | 1000 | 100   | 200.7±1.20µs      | 205.2±3.24µs      | 4.9 Ktx/sec    | 4.8 Ktx/sec    | 
| sqlite      | 💿       | location | non_unique  | 0    | 100   | 48.6±0.22µs       | 50.2±3.32µs       | 20.1 Ktx/sec   | 19.4 Ktx/sec   | 
| sqlite      | 💿       | location | non_unique  | 1000 | 100   | 52.4±0.21µs       | 52.0±0.25µs       | 18.6 Ktx/sec   | 18.8 Ktx/sec   | 
| sqlite      | 💿       | location | unique      | 0    | 100   | 51.0±7.02µs       | 50.7±2.72µs       | 19.1 Ktx/sec   | 19.2 Ktx/sec   | 
| sqlite      | 💿       | location | unique      | 1000 | 100   | 55.6±0.28µs       | 55.6±0.31µs       | 17.6 Ktx/sec   | 17.6 Ktx/sec   | 
| sqlite      | 💿       | person   | multi_index | 0    | 100   | 117.5±2.21µs      | 121.1±5.76µs      | 8.3 Ktx/sec    | 8.1 Ktx/sec    | 
| sqlite      | 💿       | person   | multi_index | 1000 | 100   | 231.0±1.97µs      | 231.9±0.35µs      | 4.2 Ktx/sec    | 4.2 Ktx/sec    | 
| sqlite      | 💿       | person   | non_unique  | 0    | 100   | 49.5±1.29µs       | 50.6±1.16µs       | 19.7 Ktx/sec   | 19.3 Ktx/sec   | 
| sqlite      | 💿       | person   | non_unique  | 1000 | 100   | 61.2±0.30µs       | 62.0±10.46µs      | 15.9 Ktx/sec   | 15.7 Ktx/sec   | 
| sqlite      | 💿       | person   | unique      | 0    | 100   | 50.2±1.54µs       | 50.7±0.33µs       | 19.5 Ktx/sec   | 19.3 Ktx/sec   | 
| sqlite      | 💿       | person   | unique      | 1000 | 100   | 56.1±0.59µs       | 56.3±0.56µs       | 17.4 Ktx/sec   | 17.3 Ktx/sec   | 
| sqlite      | 🧠       | location | multi_index | 0    | 100   | 117.8±0.28µs      | 120.3±0.27µs      | 8.3 Ktx/sec    | 8.1 Ktx/sec    | 
| sqlite      | 🧠       | location | multi_index | 1000 | 100   | 169.8±0.32µs      | 169.8±0.42µs      | 5.8 Ktx/sec    | 5.8 Ktx/sec    | 
| sqlite      | 🧠       | location | non_unique  | 0    | 100   | 43.3±0.32µs       | 43.2±0.46µs       | 22.6 Ktx/sec   | 22.6 Ktx/sec   | 
| sqlite      | 🧠       | location | non_unique  | 1000 | 100   | 44.8±0.40µs       | 43.4±0.26µs       | 21.8 Ktx/sec   | 22.5 Ktx/sec   | 
| sqlite      | 🧠       | location | unique      | 0    | 100   | 44.4±0.30µs       | 44.3±0.49µs       | 22.0 Ktx/sec   | 22.0 Ktx/sec   | 
| sqlite      | 🧠       | location | unique      | 1000 | 100   | 48.2±0.27µs       | 47.0±0.26µs       | 20.3 Ktx/sec   | 20.8 Ktx/sec   | 
| sqlite      | 🧠       | person   | multi_index | 0    | 100   | 106.6±0.39µs      | 109.7±0.47µs      | 9.2 Ktx/sec    | 8.9 Ktx/sec    | 
| sqlite      | 🧠       | person   | multi_index | 1000 | 100   | 189.2±0.37µs      | 190.9±0.29µs      | 5.2 Ktx/sec    | 5.1 Ktx/sec    | 
| sqlite      | 🧠       | person   | non_unique  | 0    | 100   | 43.8±0.35µs       | 44.1±0.31µs       | 22.3 Ktx/sec   | 22.2 Ktx/sec   | 
| sqlite      | 🧠       | person   | non_unique  | 1000 | 100   | 47.2±0.25µs       | 48.1±0.30µs       | 20.7 Ktx/sec   | 20.3 Ktx/sec   | 
| sqlite      | 🧠       | person   | unique      | 0    | 100   | 44.8±0.26µs       | 44.8±0.31µs       | 21.8 Ktx/sec   | 21.8 Ktx/sec   | 
| sqlite      | 🧠       | person   | unique      | 1000 | 100   | 49.0±0.32µs       | 48.4±0.30µs       | 19.9 Ktx/sec   | 20.2 Ktx/sec   | 
| stdb_module | 💿       | location | multi_index | 0    | 100   | 658.9±73.14µs     | 802.0±33.36µs     | 1517 tx/sec    | 1246 tx/sec    | 
| stdb_module | 💿       | location | multi_index | 1000 | 100   | 804.1±89.84µs     | 1066.7±8.03µs     | 1243 tx/sec    | 937 tx/sec     | 
| stdb_module | 💿       | location | non_unique  | 0    | 100   | 438.9±37.54µs     | 426.9±93.74µs     | 2.2 Ktx/sec    | 2.3 Ktx/sec    | 
| stdb_module | 💿       | location | non_unique  | 1000 | 100   | 537.3±69.30µs     | 762.7±66.37µs     | 1861 tx/sec    | 1311 tx/sec    | 
| stdb_module | 💿       | location | unique      | 0    | 100   | 620.9±10.78µs     | 678.6±8.02µs      | 1610 tx/sec    | 1473 tx/sec    | 
| stdb_module | 💿       | location | unique      | 1000 | 100   | 635.6±93.64µs     | 917.8±93.75µs     | 1573 tx/sec    | 1089 tx/sec    | 
| stdb_module | 💿       | person   | multi_index | 0    | 100   | 811.6±198.57µs    | 1520.0±19.35µs    | 1232 tx/sec    | 657 tx/sec     | 
| stdb_module | 💿       | person   | multi_index | 1000 | 100   | 869.1±19.13µs     | 1493.1±75.06µs    | 1150 tx/sec    | 669 tx/sec     | 
| stdb_module | 💿       | person   | non_unique  | 0    | 100   | 557.5±25.01µs     | 636.6±28.35µs     | 1793 tx/sec    | 1570 tx/sec    | 
| stdb_module | 💿       | person   | non_unique  | 1000 | 100   | 701.5±127.26µs    | 850.9±23.67µs     | 1425 tx/sec    | 1175 tx/sec    | 
| stdb_module | 💿       | person   | unique      | 0    | 100   | 829.1±68.87µs     | 939.8±26.00µs     | 1206 tx/sec    | 1064 tx/sec    | 
| stdb_module | 💿       | person   | unique      | 1000 | 100   | 840.8±71.60µs     | 1398.5±87.23µs    | 1189 tx/sec    | 715 tx/sec     | 
| stdb_module | 🧠       | location | multi_index | 0    | 100   | 683.2±2.19µs      | 718.2±4.41µs      | 1463 tx/sec    | 1392 tx/sec    | 
| stdb_module | 🧠       | location | multi_index | 1000 | 100   | 715.7±13.85µs     | 921.4±5.43µs      | 1397 tx/sec    | 1085 tx/sec    | 
| stdb_module | 🧠       | location | non_unique  | 0    | 100   | 403.9±2.14µs      | 420.1±2.15µs      | 2.4 Ktx/sec    | 2.3 Ktx/sec    | 
| stdb_module | 🧠       | location | non_unique  | 1000 | 100   | 447.4±27.83µs     | 635.7±19.99µs     | 2.2 Ktx/sec    | 1573 tx/sec    | 
| stdb_module | 🧠       | location | unique      | 0    | 100   | 434.1±86.86µs     | 558.2±1.47µs      | 2.2 Ktx/sec    | 1791 tx/sec    | 
| stdb_module | 🧠       | location | unique      | 1000 | 100   | 622.6±6.64µs      | 742.0±19.45µs     | 1606 tx/sec    | 1347 tx/sec    | 
| stdb_module | 🧠       | person   | multi_index | 0    | 100   | 777.7±1.77µs      | 826.2±5.98µs      | 1285 tx/sec    | 1210 tx/sec    | 
| stdb_module | 🧠       | person   | multi_index | 1000 | 100   | 1450.7±110.70µs   | 1141.4±151.91µs   | 689 tx/sec     | 876 tx/sec     | 
| stdb_module | 🧠       | person   | non_unique  | 0    | 100   | 538.6±8.41µs      | 420.1±98.20µs     | 1856 tx/sec    | 2.3 Ktx/sec    | 
| stdb_module | 🧠       | person   | non_unique  | 1000 | 100   | 793.3±12.66µs     | 790.0±12.72µs     | 1260 tx/sec    | 1265 tx/sec    | 
| stdb_module | 🧠       | person   | unique      | 0    | 100   | 727.2±68.25µs     | 845.3±74.40µs     | 1375 tx/sec    | 1183 tx/sec    | 
| stdb_module | 🧠       | person   | unique      | 1000 | 100   | 1097.1±23.77µs    | 1152.4±14.87µs    | 911 tx/sec     | 867 tx/sec     | 
| stdb_raw    | 💿       | location | multi_index | 0    | 100   | 277.4±1.03µs      | 304.8±0.38µs      | 3.5 Ktx/sec    | 3.2 Ktx/sec    | 
| stdb_raw    | 💿       | location | multi_index | 1000 | 100   | 332.9±238.34µs    | 366.9±82.40µs     | 2.9 Ktx/sec    | 2.7 Ktx/sec    | 
| stdb_raw    | 💿       | location | non_unique  | 0    | 100   | 129.8±0.12µs      | 136.9±0.17µs      | 7.5 Ktx/sec    | 7.1 Ktx/sec    | 
| stdb_raw    | 💿       | location | non_unique  | 1000 | 100   | 136.4±0.96µs      | 171.2±110.79µs    | 7.2 Ktx/sec    | 5.7 Ktx/sec    | 
| stdb_raw    | 💿       | location | unique      | 0    | 100   | 201.3±10.79µs     | 214.4±0.33µs      | 4.9 Ktx/sec    | 4.6 Ktx/sec    | 
| stdb_raw    | 💿       | location | unique      | 1000 | 100   | 243.8±164.83µs    | 267.2±58.35µs     | 4.0 Ktx/sec    | 3.7 Ktx/sec    | 
| stdb_raw    | 💿       | person   | multi_index | 0    | 100   | 576.4±0.86µs      | 645.3±0.37µs      | 1734 tx/sec    | 1549 tx/sec    | 
| stdb_raw    | 💿       | person   | multi_index | 1000 | 100   | 615.2±0.88µs      | 698.6±2.60µs      | 1625 tx/sec    | 1431 tx/sec    | 
| stdb_raw    | 💿       | person   | non_unique  | 0    | 100   | 181.5±11.55µs     | 195.0±0.24µs      | 5.4 Ktx/sec    | 5.0 Ktx/sec    | 
| stdb_raw    | 💿       | person   | non_unique  | 1000 | 100   | 208.7±166.01µs    | 233.8±161.81µs    | 4.7 Ktx/sec    | 4.2 Ktx/sec    | 
…